test(red): 크루 출석 기록 추가 테스트
